Set qos on subscriptions to 2.
As per https://stackoverflow.com/questions/33480730/understanding-mqtt-subscriber-qos
I believe this is the maximum qos that we can receive.

So we were forcing all incoming messages to be qos 0 regardless of what
the publisher requested.

I believe if the client is using qos=1 it will still use qos=1, but am a
still a bit unclear on this.